Python layout changes.
Install capisuite python module in pkgpython_PYTHON, this puts capisuite as
a module into the standard python path.

cs_helpers.py(in) was renamed to helpers.py(in) and also put into
the capisuite module. The old autoconf code put cs_helpers into
site-packages, and that is not part of the default sys.path any more.
That may be a Debian issue, but having the capisuite python code in two
different directories seemed undesirable enough to warrant this change.