don't return empty user-values, but return default instead
since otherwise it contradicts the documentation when e.g. sending mail
notifications, as the code that requests e.g. the fax.email property
provides the user as default.
But when using the empty string as documented to use the user's name as
sender, it returns the empty string and that then causes sending the
mail to fail, as there is no recipient.