use tiff2pdf instead of converting first to ps and then to pdf
this is not only a lot faster, but also allows lossless conversion.
Along that line, don't try to alter the pagesize during conversion to
PDF. PDF-viewer applications do a great job adjusting to the paper-size
as the user wants it, so let the user make the decision. After all when
having the PDF, it is unlikely that it will be printed out anyway.